Fix handling of audio delay
This patch renames `audio_offset` to `audio_delay`, to match VLC.
Previously, the code incorrectly added the offset, and now it properly
subtracts the delay.

This means that all the audio offsets specified in the existing lec*.py
files should be updated; that will be fixed in a future commit.